add sample controller
