+++ incomingcallsample/readm.MD
@@ -1,0 +1,5 @@
+# Start locally
+```dotnetcli
+mvn clean package
+java -jar .\target\incomingcallsample-0.0.1-SNAPSHOT.ja
+```
