--- incomingcallsample/src/main/java/com/communication/incomingcallsample/Greeting.java
@@ -1,20 +1,0 @@
-package com.communication.incomingcallsample;
-
-public class Greeting {
-
-	private final long id;
-	private final String content;
-
-	public Greeting(long id, String content) {
-		this.id = id;
-		this.content = content;
-	}
-
-	public long getId() {
-		return id;
-	}
-
-	public String getContent() {
-		return content;
-	}
-}
