+++ incomingcallsample/src/main/java/com/communication/incomingcallsample/IncomingcallsampleApplication.java
@@ -2,12 +2,25 @@ package com.communication.incomingcallsample;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.context.annotation.ComponentScan;
+import java.util.*;
+
+import com.communication.incomingcallsample.Controller.IncomingCallController;
+
 
 @SpringBootApplication
+@ComponentScan(basePackageClasses= IncomingCallController.class)
 public class IncomingcallsampleApplication {
+	final static String url = "http://localhost:9008";
+    final static String serverPort = "9008";
 
 	public static void main(String[] args) {
-		SpringApplication.run(IncomingcallsampleApplication.class, args);
+		SpringApplication app = new SpringApplication(IncomingcallsampleApplication.class);
+        app.setDefaultProperties(Collections
+          .singletonMap("server.port", serverPort));
+        app.run(args);
+
+
 	}
 
 }
